update function structure
