+++ tests/test_onnx/__init__.py
@@ -1,3 +1,6 @@
-from .utils import convert_result_list, verify_model
+from .utils import (WrapFunction, convert_result_list, ort_validate,
+                    verify_model)
 
-__all__ = ['verify_model', 'convert_result_list']
+__all__ = [
+    'WrapFunction', 'verify_model', 'convert_result_list', 'ort_validate'
+]
