+++ tests/test_onnx/utils.py
@@ -1,9 +1,11 @@
 import warnings
 from os import path as osp
 
+import numpy as np
 import onnx
 import onnxruntime as ort
 import torch
+import torch.nn as nn
 
 ort_custom_op_path = ''
 try:
@@ -12,6 +14,50 @@ try:
 except (ImportError, ModuleNotFoundError):
     warnings.warn('If input model has custom op from mmcv, \
         you may have to build mmcv with ONNXRuntime from source.')
+
+
+class WrapFunction(nn.Module):
+    """Wrap the function to be tested for torch.onnx.export tracking."""
+
+    def __init__(self, wrapped_function):
+        super(WrapFunction, self).__init__()
+        self.wrapped_function = wrapped_function
+
+    def forward(self, *args, **kwargs):
+        return self.wrapped_function(*args, **kwargs)
+
+
+def ort_validate(model, feats, onnx_io='tmp.onnx'):
+    """Validate the output of the onnxruntime backend is the same as the output
+    generated by torch.
+
+    Args:
+        model (nn.Module): the model to be verified
+        feats (list(torch.Tensor) | torch.Tensor): the input of model
+        onnx_io (str): the name of onnx output file
+    """
+    with torch.no_grad():
+        torch.onnx.export(
+            model,
+            feats,
+            onnx_io,
+            export_params=True,
+            keep_initializers_as_inputs=True,
+            do_constant_folding=True,
+            verbose=False,
+            opset_version=11)
+
+    onnx_outputs = verify_model(feats)
+
+    torch_outputs = convert_result_list(model.forward(feats))
+    torch_outputs = [
+        torch_output.detach().numpy() for torch_output in torch_outputs
+    ]
+
+    # match torch_outputs and onnx_outputs
+    for i in range(len(onnx_outputs)):
+        np.testing.assert_allclose(
+            torch_outputs[i], onnx_outputs[i], rtol=1e-03, atol=1e-05)
 
 
 def verify_model(feat, onnx_io='tmp.onnx'):
